requirements.txt
Pillow
cryptography
pyqt6
